# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PlaylistForge REST API",
    "version": "1.0.0",
    "description": "API documentation for PlaylistForge MERN music playlist creator with Spotify Web API & Redis integration."
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development Server"
    }
  ],
  "paths": {
...
    "/search": {
      "get": {
        "summary": "Search tracks with 300s Redis cache",
        "parameters": [
          { "name": "q", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Returns search results" } }
      }
    },
    "/playlists": {
      "get": {
        "summary": "Get user owned playlists",
        "responses": { "200": { "description": "Array of playlists" } }
      },
      "post": {
        "summary": "Create new playlist",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "isPublic": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Playlist created" } }
      }
    },
...
  }
}